+++ RealmSwift/Tests/MigrationTests.swift
@@ -22,7 +22,7 @@ import Realm
 import Realm.Private
 import Foundation
 
-private func realmWithCustomSchema(path: String, schema :RLMSchema) -> RLMRealm {
+private func realmWithCustomSchema(path: String, schema: RLMSchema) -> RLMRealm {
     return RLMRealm(path: path, key: nil, readOnly: false, inMemory: false, dynamic: true, schema: schema, error: nil)!
 }
 
@@ -32,14 +32,14 @@ private func realmWithSingleClass(path: String, objectSchema: RLMObjectSchema) -
     return realmWithCustomSchema(path, schema)
 }
 
-private func realmWithSingleClassProperties(path: String, className: String, properties: [AnyObject]) -> RLMRealm {
+internal func realmWithSingleClassProperties(path: String, className: String, properties: [AnyObject]) -> RLMRealm {
     let objectSchema = RLMObjectSchema(className: className, objectClass: MigrationObject.self, properties: properties)
     return realmWithSingleClass(path, objectSchema)
 }
 
 class MigrationTests: TestCase {
 
-    // MARK Utility methods
+    // MARK: Utility methods
 
     // create realm at path and test version is 0
     private func createAndTestRealmAtPath(realmPath: String) {
@@ -51,7 +51,9 @@ class MigrationTests: TestCase {
     }
 
     // migrate realm at path and ensure migration
-    private func migrateAndTestRealm(realmPath: String, shouldRun: Bool = true, schemaVersion: UInt = 1, autoMigration: Bool = false, block: MigrationBlock? = nil) {
+    private func migrateAndTestRealm(realmPath: String, isNecessary: Bool = false, shouldRun: Bool = true, schemaVersion: UInt = 1, autoMigration: Bool = false, block: MigrationBlock? = nil) {
+        var error: NSError? = nil
+        XCTAssertEqual(Realm.migrationRequired(atPath: realmPath, error: &error), isNecessary)
         var didRun = false
         setSchemaVersion(schemaVersion, realmPath, { migration, oldSchemaVersion in
             if let block = block {
@@ -69,9 +71,10 @@ class MigrationTests: TestCase {
         }
 
         XCTAssertEqual(didRun, shouldRun)
+        XCTAssertFalse(Realm.migrationRequired(atPath: realmPath))
     }
 
-    // MARK Test cases
+    // MARK: Test cases
 
     func testSetDefaultRealmSchemaVersion() {
         createAndTestRealmAtPath(Realm.defaultPath)
@@ -117,13 +120,12 @@ class MigrationTests: TestCase {
     }
 
     func testMigrationProperties() {
-        let prop = RLMProperty(name: "stringCol", type: RLMPropertyType.Int, objectClassName: nil, indexed: false)
-        autoreleasepool { () -> () in
+        autoreleasepool {
+            let prop = RLMProperty(name: "stringCol", type: .Int, objectClassName: nil, indexed: false)
             realmWithSingleClassProperties(Realm.defaultPath, "SwiftStringObject", [prop])
-            return
         }
 
-        migrateAndTestRealm(Realm.defaultPath, block: { migration, oldSchemaVersion in
+        migrateAndTestRealm(Realm.defaultPath, isNecessary: true, block: { migration, oldSchemaVersion in
             XCTAssertEqual(migration.oldSchema.objectSchema.count, 1)
             XCTAssertGreaterThan(migration.newSchema.objectSchema.count, 1)
             XCTAssertEqual(migration.oldSchema.objectSchema[0].properties.count, 1)
@@ -330,4 +332,3 @@ class MigrationTests: TestCase {
         XCTAssertEqual(Realm().objects(SwiftBoolObject).count, 4)
     }
 }
-