+++ RealmSwift/Tests/RealmTests.swift
@@ -17,6 +17,7 @@
 ////////////////////////////////////////////////////////////////////////////
 
 import XCTest
+import Realm.Private
 import RealmSwift
 import Foundation
 
@@ -462,6 +463,20 @@ class RealmTests: TestCase {
         Realm.setEncryptionKey(NSMutableData(length: 64))
         Realm.setEncryptionKey(nil, forPath: Realm.defaultPath)
         XCTAssert(true, "setting those keys should not throw")
+    }
+
+    func testMigrationRequired() {
+        let path1 = realmPathForFile("1")
+        autoreleasepool {
+            _ = Realm(path: path1)
+        }
+        XCTAssertFalse(Realm.migrationRequired(atPath: path1), "should return false immediately after creating a realm.")
+        let path2 = realmPathForFile("2")
+        autoreleasepool {
+            let prop = RLMProperty(name: "stringCol", type: .Int, objectClassName: nil, indexed: false)
+            realmWithSingleClassProperties(path2, "SwiftStringObject", [prop])
+        }
+        XCTAssert(Realm.migrationRequired(atPath: path2), "should return true immediately after creating a realm with a custom schema.")
     }
 
     func testEquals() {